fix unresponsiveness after unfocus

+++ SFMLGame/FieldOfVision.cpp
@@ -159,7 +159,7 @@ std::vector<sf::Vector2i> generate_line_of_sight(Tilemap &map, sf::Vector2i src,
 			break;
 		}
 	}
-	if (rval.size() > radius) {
+	if (rval.size() > (size_t)radius) {
 		rval.clear();
 	}
 	return rval;
